Added template
